Refactored many api functions to httputil/api package from lfs
Also moved ntlm to httputil since too http-specific to be contained in auth
