+++ httputil/ntlm.go
@@ -1,4 +1,4 @@
-package auth
+package httputil
 
 import (
 	"bytes"
@@ -12,12 +12,12 @@ import (
 	"strings"
 	"sync/atomic"
 
+	"github.com/github/git-lfs/auth"
 	"github.com/github/git-lfs/config"
-	"github.com/github/git-lfs/httputil"
 	"github.com/github/git-lfs/vendor/_nuts/github.com/ThomsonReutersEikon/go-ntlm/ntlm"
 )
 
-func ntlmClientSession(c *config.Configuration, creds Creds) (ntlm.ClientSession, error) {
+func ntlmClientSession(c *config.Configuration, creds auth.Creds) (ntlm.ClientSession, error) {
 	if c.NtlmSession != nil {
 		return c.NtlmSession, nil
 	}
@@ -39,13 +39,13 @@ func ntlmClientSession(c *config.Configuration, creds Creds) (ntlm.ClientSession
 	return session, nil
 }
 
-func DoNTLMRequest(request *http.Request, retry bool) (*http.Response, error) {
+func doNTLMRequest(request *http.Request, retry bool) (*http.Response, error) {
 	handReq, err := cloneRequest(request)
 	if err != nil {
 		return nil, err
 	}
 
-	res, err := httputil.NewHttpClient(config.Config, handReq.Host).Do(handReq)
+	res, err := NewHttpClient(config.Config, handReq.Host).Do(handReq)
 	if err != nil && res == nil {
 		return nil, err
 	}
@@ -53,7 +53,7 @@ func DoNTLMRequest(request *http.Request, retry bool) (*http.Response, error) {
 	//If the status is 401 then we need to re-authenticate, otherwise it was successful
 	if res.StatusCode == 401 {
 
-		creds, err := GetCreds(request)
+		creds, err := auth.GetCreds(request)
 		if err != nil {
 			return nil, err
 		}
@@ -80,10 +80,10 @@ func DoNTLMRequest(request *http.Request, retry bool) (*http.Response, error) {
 
 		//If the status is 401 then we need to re-authenticate
 		if res.StatusCode == 401 && retry == true {
-			return DoNTLMRequest(challengeReq, false)
+			return doNTLMRequest(challengeReq, false)
 		}
 
-		SaveCredentials(creds, res)
+		auth.SaveCredentials(creds, res)
 
 		return res, nil
 	}
@@ -92,7 +92,7 @@ func DoNTLMRequest(request *http.Request, retry bool) (*http.Response, error) {
 
 func negotiate(request *http.Request, message string) ([]byte, error) {
 	request.Header.Add("Authorization", message)
-	res, err := httputil.NewHttpClient(config.Config, request.Host).Do(request)
+	res, err := NewHttpClient(config.Config, request.Host).Do(request)
 
 	if res == nil && err != nil {
 		return nil, err
@@ -109,7 +109,7 @@ func negotiate(request *http.Request, message string) ([]byte, error) {
 	return ret, nil
 }
 
-func challenge(request *http.Request, challengeBytes []byte, creds Creds) (*http.Response, error) {
+func challenge(request *http.Request, challengeBytes []byte, creds auth.Creds) (*http.Response, error) {
 	challenge, err := ntlm.ParseChallengeMessage(challengeBytes)
 	if err != nil {
 		return nil, err
@@ -128,7 +128,7 @@ func challenge(request *http.Request, challengeBytes []byte, creds Creds) (*http
 
 	authMsg := base64.StdEncoding.EncodeToString(authenticate.Bytes())
 	request.Header.Add("Authorization", "NTLM "+authMsg)
-	return httputil.NewHttpClient(config.Config, request.Host).Do(request)
+	return NewHttpClient(config.Config, request.Host).Do(request)
 }
 
 func parseChallengeResponse(response *http.Response) ([]byte, error) {
