+++ httputil/request_error_test.go
@@ -1,4 +1,4 @@
-package lfs
+package httputil
 
 import (
 	"bytes"
